backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
numpy==2.3.2
pillow==11.3.0
python-multipart==0.0.20
